requirements.txt
amqp==2.6.1
asgiref==3.3.4
attrs==20.3.0
autopep8==1.5.7
bcrypt==3.2.0
billiard==3.6.4.0
blessed==1.18.0
boto3==1.17.84
botocore==1.20.84
cached-property==1.5.2
celery==4.4.7
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.5.0
dj-database-url==0.5.0
Django==3.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-import-export==2.5.0
django-on-heroku==1.1.2
django-phone-field==1.8.1
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docker==5.0.0
docker-compose==1.29.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
future==0.16.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
kombu==4.6.11
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.9
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.1
semantic-version==2.8.5
simplejson==3.17.2
six==1.14.0
sqlparse==0.4.1
tablib==3.0.0
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
wcwidth==0.1.9
websocket-client==0.58.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
